Refactored Database interaction functionality

+++ src/database/Database.cpp
@@ -35,25 +35,43 @@ class PrepareStatement{
 
 };
 
+// sqlite3* Database::getDatabase(){
+//     return database;
+// }
+
+
 bool Database::fileExists(const char* fileName) {
     return std::filesystem::exists(fileName);
 }
 
 Database::Database(const char* file_name) : database(nullptr){
     
-    if (!fileExists(file_name)) {
-        throw std::runtime_error("Database file does not exist: " + std::string(file_name));
+    // If we are opening an in-memory database (during unit tests)
+    if (std::string(file_name) == ":memory:") {
+        
+        int rc = sqlite3_open(file_name, &database);
+        if (rc != SQLITE_OK) {
+            std::string err_msg = "Failed to open in-memory DB: " + std::string(sqlite3_errmsg(database));
+            sqlite3_close(database);
+            throw std::runtime_error(err_msg);
+        }
     }
-    
-    int rc = sqlite3_open(file_name, &database);
 
-    if (rc != SQLITE_OK) {
-        std::string err_msg = "Failed to open DB" + std::string(sqlite3_errmsg(database));
-        sqlite3_close(database);
-        throw std::runtime_error(err_msg);
-    } 
-    else {
-        std::cout << "Opened DB\n";
+    else{    
+        if (!fileExists(file_name)) {
+            throw std::runtime_error("Database file does not exist: " + std::string(file_name));
+        }
+        
+        int rc = sqlite3_open(file_name, &database);
+
+        if (rc != SQLITE_OK) {
+            std::string err_msg = "Failed to open DB" + std::string(sqlite3_errmsg(database));
+            sqlite3_close(database);
+            throw std::runtime_error(err_msg);
+        } 
+        else {
+            std::cout << "Opened DB\n";
+        }
     }
 }
 
@@ -125,9 +143,8 @@ std::vector<Item> Database::loadUserCart(int user_id) {
     return user_cart;
 }
 
-void Database::addToCart(Product product, int user_id) {
+void Database::addToCart(int product_id, int user_id) {
 
-    int product_id = product.getProductID();
     
     const char* sql_command =  "SELECT COUNT(*) " 
                                "FROM USER_CART uc "
@@ -149,7 +166,7 @@ void Database::addToCart(Product product, int user_id) {
 
         if(count > 0){
             
-            std::cout << "Product Exists, Updating Count" << std::endl;
+            std::cout << "Product count in Cart: " << count << ", Incrementing count to: " << count << std::endl;
 
             const char* sql_command = "UPDATE CART_ITEMS " 
                                      "SET quantity = quantity + 1 " 
@@ -171,13 +188,13 @@ void Database::addToCart(Product product, int user_id) {
                 throw std::runtime_error(sqlite3_errmsg(database));
             }
             else{
-                std::cout << "Count Updated Successfully";
+                std::cout << "Count Increased Successfully" << std::endl;
             }
         }
 
         else{
             
-            std::cout << "Product Does Not Exist, Adding to Table" << std::endl;
+            std::cout << "Product Not Yet in Cart, Adding to Table" << std::endl;
 
             const char* sql_command = "INSERT INTO CART_ITEMS (fk_cart_id, fk_product_id, quantity) "
                             "VALUES("
@@ -199,42 +216,94 @@ void Database::addToCart(Product product, int user_id) {
                 throw std::runtime_error(sqlite3_errmsg(database));
             }
             else{
-                std::cout << "Product Inserted Successfully!";
+                std::cout << "Product Inserted Successfully!" <<std::endl;
             }
         }        
     }          
 }
 
 void Database::removeFromCart(int user_id, int product_id) {
+
+    const char* sql_command =  "SELECT quantity " 
+                            "FROM USER_CART uc "
+                            "JOIN CART_ITEMS ci ON uc.CART_ID = ci.fk_cart_id "
+                            "WHERE uc.fk_user_id = ? "
+                            "AND ci.fk_product_id = ?";
+
+    PrepareStatement stmt_query(sql_command, database);
+
+    // Bind user_id to the first placeholder
+    sqlite3_bind_int(stmt_query.get_stmt(), 1, user_id);
+
+    // Bind product_id to the second placeholder
+    sqlite3_bind_int(stmt_query.get_stmt(), 2, product_id);
+
+    if(sqlite3_step(stmt_query.get_stmt()) == SQLITE_ROW){
+        
+        int count = sqlite3_column_int(stmt_query.get_stmt(), 0);
+
+        if(count >= 2){
+            
+            std::cout << "Product count in Cart: " << count << ", Decrementing count to: " << count - 1 << std::endl;
+
+            const char* sql_command = "UPDATE CART_ITEMS " 
+                                     "SET quantity = quantity - 1 " 
+                                     "WHERE fk_cart_id = "
+                                     "(SELECT CART_ID FROM USER_CART "
+                                     "WHERE fk_user_id = ?) " 
+                                     "AND fk_product_id = ?;";
+            
+            PrepareStatement stmt_update(sql_command, database);
+            
+            // Bind user_id to the second placeholder
+            sqlite3_bind_int(stmt_update.get_stmt(), 1, user_id);
     
-    const char* sql_command = "DELETE FROM CART_ITEMS "
+            // Bind product_id to the second placeholder
+            sqlite3_bind_int(stmt_update.get_stmt(), 2, product_id);
+   
+            int rc = sqlite3_step(stmt_update.get_stmt());
+            if (rc != SQLITE_DONE) {
+                throw std::runtime_error(sqlite3_errmsg(database));
+            }
+            else{
+                std::cout << "Count Decreased Successfully" << std::endl;
+            }
+        }
+
+        else{
+
+            std::cout << "Product count in Cart: " << count << ", Deleting Item from Cart" << std::endl;
+            
+             const char* sql_command = "DELETE FROM CART_ITEMS "
                               "WHERE fk_cart_id = (SELECT CART_ID FROM USER_CART WHERE fk_user_id = ?) "
                               "AND fk_product_id = ?";
     
-    PrepareStatement stmt_delete(sql_command, database);
+            PrepareStatement stmt_delete(sql_command, database);
 
-    sqlite3_bind_int(stmt_delete.get_stmt(), 1, user_id);
-    sqlite3_bind_int(stmt_delete.get_stmt(), 2, product_id);
+            sqlite3_bind_int(stmt_delete.get_stmt(), 1, user_id);
+            sqlite3_bind_int(stmt_delete.get_stmt(), 2, product_id);
 
-    int rc = sqlite3_step(stmt_delete.get_stmt());
-    if (rc != SQLITE_DONE) {
-        throw std::runtime_error(sqlite3_errmsg(database));
-    }
-    else{
-        std::cout << "Cart Item Deleted Successfully!";
-    }
+            int rc = sqlite3_step(stmt_delete.get_stmt());
+            if (rc != SQLITE_DONE) {
+                throw std::runtime_error(sqlite3_errmsg(database));
+            }
+            else{
+                std::cout << "Cart Item Deleted Successfully!" << std::endl;
+            }
+        }        
+    }          
 }
 
 std::string Database::fetchPassword(std::string username){
     
-    const char* sql_command = "SELECT * FROM USER_INFORMATION WHERE username = ?";
+    const char* sql_command = "SELECT password FROM USER_INFORMATION WHERE username = ?";
     PrepareStatement stmt_insert(sql_command, database);
 
     sqlite3_bind_text(stmt_insert.get_stmt(), 1, username.c_str(), -1, SQLITE_STATIC);
     
     if(sqlite3_step(stmt_insert.get_stmt()) == SQLITE_ROW){
         
-        const unsigned char* password_char = sqlite3_column_text(stmt_insert.get_stmt(), 2);
+        const unsigned char* password_char = sqlite3_column_text(stmt_insert.get_stmt(), 0);
         std::string password(reinterpret_cast<const char*>(password_char));
         std::cout<<"Password: "<<password<<std::endl;
         return password;
@@ -255,6 +324,6 @@ void Database::executeQuery(const char* sql_query){
     }
     else{
         std::string query_str(sql_query);
-        std::cout << "Executed Statement Successfully: " + query_str << std::endl;
+        // std::cout << "Executed Statement Successfully: " + query_str << std::endl;
     }
 }
